Add UI dialog and simplify refactoring provider
