+++ src/Features/CSharp/Portable/CodeRefactorings/PushMemberUp/PushMemberUpCodeRefactoringProvider.cs
@@ -2,15 +2,88 @@
 
 using System.Collections.Generic;
 using System.Composition;
+using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.CodeRefactorings;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
 namespace Microsoft.CodeAnalysis.CSharp.CodeRefactorings.PushMemberUp
 {
     [ExportCodeRefactoringProvider(LanguageNames.CSharp, Name = nameof(PredefinedCodeRefactoringProviderNames.PushMember)), Shared]
-    internal class PushMemberUpCodeRefactoringProvider : AbstractPushMemberUpCodeRefactoringProvider
+    internal class PushMemberUpCodeRefactoringProvider : CodeRefactoringProvider
     {
-        protected override bool IsUserSelectIdentifer(SyntaxNode userSelectedSyntax)
+        protected SemanticModel SemanticModel { get; set; }
+
+        protected CodeRefactoringContext Context { get; set; }
+
+        public override async Task ComputeRefactoringsAsync(CodeRefactoringContext context)
+        {
+            Context = context;
+            SemanticModel = await context.Document.GetSemanticModelAsync(context.CancellationToken).ConfigureAwait(false);
+            var root = await context.Document.GetSyntaxRootAsync(context.CancellationToken).ConfigureAwait(false);
+            var userSelectedNode = root.FindNode(context.Span);
+
+            if(!IsUserSelectIdentifer(userSelectedNode))
+            {
+                return;
+            }
+
+            var userSelectedNodeSymbol = SemanticModel.GetDeclaredSymbol(userSelectedNode);
+            if (userSelectedNodeSymbol == null)
+            {
+                return;
+            }
+
+            var allTargetClasses = FindAllTargetBaseClasses(userSelectedNodeSymbol.ContainingType);
+            var allTargetInterfaces = FindAllTargetInterfaces(userSelectedNodeSymbol.ContainingType);
+
+            if (userSelectedNodeSymbol.Kind == SymbolKind.Method ||
+                userSelectedNodeSymbol.Kind == SymbolKind.Property ||
+                userSelectedNodeSymbol.Kind == SymbolKind.Event)
+            {
+                ProcessingClassesRefactoringWithQuickAction(allTargetClasses, userSelectedNode);
+                ProcessingInterfacesRefactoringQuickAction(allTargetInterfaces, userSelectedNode);
+                ProcessingRefactoringViaDialogBox(allTargetClasses, allTargetInterfaces, userSelectedNodeSymbol.ContainingType);
+            }
+            else if (userSelectedNodeSymbol.Kind == SymbolKind.Field)
+            {
+                ProcessingClassesRefactoringWithQuickAction(allTargetClasses, userSelectedNode);
+                ProcessingRefactoringViaDialogBox(allTargetClasses, allTargetInterfaces, userSelectedNodeSymbol.ContainingType);
+            }
+        }
+
+
+        private void ProcessingRefactoringViaDialogBox(
+            IEnumerable<INamedTypeSymbol> targetClasses,
+            IEnumerable<INamedTypeSymbol> targetInterfaces,
+            INamedTypeSymbol selectedNodeOwnerSymbol)
+        {
+            var service = Context.Document.Project.Solution.Workspace.Services.GetRequiredService<IPushMemberUpService>();
+
+            var dialogAction = new PushMemberUpWithDialogCodeAction(SemanticModel, service, selectedNodeOwnerSymbol);
+            Context.RegisterRefactoring(dialogAction);
+        }
+
+        private IEnumerable<INamedTypeSymbol> FindAllTargetInterfaces(INamedTypeSymbol selectedNodeOwnerSymbol)
+        {
+            return selectedNodeOwnerSymbol.AllInterfaces.Where(eachInterface => eachInterface.DeclaringSyntaxReferences.Length > 0);
+        }
+
+        private IEnumerable<INamedTypeSymbol> FindAllTargetBaseClasses(INamedTypeSymbol selectedNodeOwnerSymbol)
+        {
+            var allBasesClasses = new List<INamedTypeSymbol>();
+            while (selectedNodeOwnerSymbol.BaseType != null)
+            {
+                if (selectedNodeOwnerSymbol.BaseType.DeclaringSyntaxReferences.Length > 0)
+                {
+                    allBasesClasses.Add(selectedNodeOwnerSymbol.BaseType);
+                }
+                selectedNodeOwnerSymbol = selectedNodeOwnerSymbol.BaseType;
+            }
+            return allBasesClasses;
+        }
+       
+        protected bool IsUserSelectIdentifer(SyntaxNode userSelectedSyntax)
         {
             var identifier = GetIdentifier(userSelectedSyntax);
             return identifier.Span.Contains(Context.Span);
@@ -33,11 +106,11 @@ namespace Microsoft.CodeAnalysis.CSharp.CodeRefactorings.PushMemberUp
             }
         }
 
-        protected override void ProcessingClassesRefactoring(IEnumerable<INamedTypeSymbol> targetClasses, SyntaxNode userSelectSyntax)
+        protected void ProcessingClassesRefactoringWithQuickAction(IEnumerable<INamedTypeSymbol> targetClasses, SyntaxNode userSelectSyntax)
         {
             foreach (var eachClass in targetClasses)
             {
-                var classPusher = new ClassPusher(eachClass, SemanticModel, userSelectSyntax, Context.Document);
+                var classPusher = new ClassPusherWithQuickAction(eachClass, SemanticModel, userSelectSyntax, Context.Document);
                 var action = classPusher.ComputeRefactoring();
                 if (action != null)
                 {
@@ -46,11 +119,11 @@ namespace Microsoft.CodeAnalysis.CSharp.CodeRefactorings.PushMemberUp
             }
         }
 
-        protected override void ProcessingInterfacesRefactoring(IEnumerable<INamedTypeSymbol> targetInterfaces, SyntaxNode userSelectSyntax)
+        protected void ProcessingInterfacesRefactoringQuickAction(IEnumerable<INamedTypeSymbol> targetInterfaces, SyntaxNode userSelectSyntax)
         {
             foreach (var eachInterface in targetInterfaces)
             {
-                var interfacePusher = new InterfacePusher(eachInterface, SemanticModel, userSelectSyntax, Context.Document);
+                var interfacePusher = new InterfacePusherWithQuickAction(eachInterface, SemanticModel, userSelectSyntax, Context.Document);
                 var action = interfacePusher.ComputeRefactoring();
 
                 if (action != null)
