+++ src/VisualStudio/Core/Def/Implementation/PushMemberUp/PushMemberUpDialogWarning.xaml
@@ -1,0 +1,104 @@
+﻿<vs:DialogWindow
+             x:Uid="PushMemberUpDialog" 
+             x:Class="Microsoft.VisualStudio.LanguageServices.Implementation.PushMemberUp.PushMemberUpDialogWarningxaml"
+             x:ClassModifier="internal"
+             x:Name="dialog"
+             xmlns:vs="clr-namespace:Microsoft.VisualStudio.PlatformUI;assembly=Microsoft.VisualStudio.Shell.15.0"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:sys="clr-namespace:System;assembly=mscorlib"
+             xmlns:self="clr-namespace:Microsoft.VisualStudio.LanguageServices.Implementation.PushMemberUp"
+             mc:Ignorable="d" 
+             SizeToContent="Height"
+             Height="290"
+             Width="480"
+             Title="Pull Members Up">
+    <Window.Resources>
+        <Thickness x:Key="okAndCancelButton">5, 3, 8, 4</Thickness>
+        <Thickness x:Key="selectDependantsButton">2, 4, 4, 2</Thickness>
+        <sys:Double x:Key="okAndCancelButtonWidth">50</sys:Double>
+        <sys:Double x:Key="okAndCancelButtonHeight">20</sys:Double>
+        <Style x:Key="RoundCornerButton" TargetType="Button">
+            <Setter Property="Template">
+                <Setter.Value>
+                    <ControlTemplate TargetType="Button">
+                        <Border CornerRadius="2" BorderThickness="1" BorderBrush="Black">
+                            <ContentPresenter x:Name="contentPresenter" ContentTemplate="{TemplateBinding ContentTemplate}" Content="{TemplateBinding Content}" HorizontalAlignment="{TemplateBinding HorizontalContentAlignment}" Margin="{TemplateBinding Padding}" VerticalAlignment="{TemplateBinding VerticalContentAlignment}"/>
+                        </Border>
+                    </ControlTemplate>
+                </Setter.Value>
+            </Setter>
+        </Style>
+        <Style x:Key="LibraryTreeViewItemStyle"
+               TargetType="{x:Type TreeViewItem}">
+            <Setter Property="IsSelected"
+                        Value="{Binding IsSelected, Mode=TwoWay}" />
+            <Setter Property="FontWeight"
+                        Value="Normal" />
+            <Style.Triggers>
+                <Trigger Property="IsSelected" Value="True">
+                    <Setter Property="IsExpanded"
+                            Value="{Binding IsExpanded, Mode=TwoWay}" />
+                    <Setter Property="FontWeight" Value="Bold" />
+                </Trigger>
+            </Style.Triggers>
+        </Style>
+    </Window.Resources>
+    <Grid Margin="10,8,10,8">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="Auto"/>
+        </Grid.RowDefinitions>
+
+        <StackPanel Grid.Row="0" Height="200" MinWidth="370">
+            <Label
+               x:Uid="SelectTargetLabel"
+               x:Name="PushUpLabel"
+               FontWeight="Bold"
+               Content="Pull Up"/>
+            <Label
+                x:Uid="Description"
+                x:Name="Description"
+                Content="  Review problems below. Click Finish to proceed pull up"/>
+            <Separator/>
+
+            <Label x:Name="TargetSelectionLabel"
+                   Content="Found Problems"/>
+        </StackPanel>
+        <Grid Grid.Row="1" Margin="5,5,5,0" Height="50">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="Auto"/>
+                <ColumnDefinition Width="Auto"/>
+            </Grid.ColumnDefinitions>
+            <StackPanel Grid.Row="3"
+                   Orientation="Vertical"
+                   MaxHeight="40" Margin="0,8,0,8" Width="445">
+                <Grid>
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition Width="*"/>
+                        <ColumnDefinition Width="Auto"/>
+                        <ColumnDefinition Width="Auto"/>
+                    </Grid.ColumnDefinitions>
+                    <Button x:Uid="OKButton"
+                    Grid.Column="1"
+                    Style="{StaticResource RoundCornerButton}" 
+                    Margin = "{StaticResource ResourceKey=okAndCancelButton}"
+                    Content="Back"
+                    Width= "{StaticResource ResourceKey=okAndCancelButtonWidth}"
+                    Height="{StaticResource ResourceKey=okAndCancelButtonHeight}"/>
+                    <Button x:Uid="CancelButton"
+                    Grid.Column="2"
+                    Style="{StaticResource RoundCornerButton}" 
+                    Margin= "{StaticResource ResourceKey=okAndCancelButton}"
+                    Content="Finish"
+                    Width= "{StaticResource ResourceKey=okAndCancelButtonWidth}"
+                    Height="{StaticResource ResourceKey=okAndCancelButtonHeight}"/>
+                </Grid>
+            </StackPanel>
+        </Grid>
+    </Grid>
+</vs:DialogWindow>
