+++ src/VisualStudio/Core/Def/Microsoft.VisualStudio.LanguageServices.csproj
@@ -50,6 +50,12 @@
     <Compile Update="Implementation\PickMembers\PickMembersDialog.xaml.cs">
       <DependentUpon>PickMembersDialog.xaml</DependentUpon>
     </Compile>
+    <Compile Update="Implementation\PushMemberUp\PushMemberUpDialog.xaml.cs">
+      <DependentUpon>PushMemberUpDialog.xaml</DependentUpon>
+    </Compile>
+    <Compile Update="Implementation\PushMemberUp\PushMemberUpDialogWarning.xaml.cs">
+      <DependentUpon>PushMemberUpDialogWarning.xaml</DependentUpon>
+    </Compile>
     <Compile Update="Implementation\Workspace\DetailedErrorInfoDialog.xaml.cs">
       <DependentUpon>DetailedErrorInfoDialog.xaml</DependentUpon>
     </Compile>
@@ -229,6 +235,12 @@
     </Page>
     <Page Include="Implementation\GenerateType\GenerateTypeDialog.xaml">
       <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="Implementation\PushMemberUp\PushMemberUpDialog.xaml">
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="Implementation\PushMemberUp\PushMemberUpDialogWarning.xaml">
       <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Include="Implementation\Workspace\DetailedErrorInfoDialog.xaml">
